requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1.7
django-admin-generator==2.6.0
django-cors-headers==4.7.0
django-redis==5.4.0
django-utils2==3.0.2
djangorestframework==3.16.0
filelock==3.18.0
fsspec==2025.3.1
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
torch==2.6.0+cpu
sentence-transformers==4.0.1
six==1.17.0
sqlparse==0.5.3
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.50.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0